Issue 4328: Add means to display objects accessible from a grob
A convenient way to display the grobs and grob-objects pointed to by
various internal properties of a given grob would be very helpful for
debugging and development purposes.  For example, it would facilitate the
creation of advanced tweaks, which frequently require "lateral" access
to other grobs.

The output of the function 'grob::display-objects' shows all of the grobs
accessible to a given grob through ly:grob-object together with the relevant
interfaces and properties.  It includes properties which are empty: not all
properties within an interface may be set for or used by a grob supporting
that interface.