Issue 4360: Reorganize smob initialization to make it more reliable
This change is a nuisance in that it requires an explicit call to
ADD_SMOB_INIT for each Scheme-related class.

However, this setup guarantees that at the point Scheme execution
commences, all predicates of the likes of "ly:grob?" will be present.
It also guarantees that if a particular type should seem required during
_other_ initializations that happen to depend on the type and get
executed earlier, it will be provided early.