Issue 4364: Allow ImageMagick's compare to exit with status 1
Apparently Ubuntu 15.04 has a version of "compare" that cannot be easily
persuaded to return anything but exit status 1 (which indicates
dissimilar images but no actual error condition) so we allow this in
script/build/output-distance.py in order to keep "make check" from
failing.

This patch is somewhat artless but does the trick.