Issue 4376: Add LY_ASSERT_DERIVED_SMOB macro
When LY_ASSERT_SMOB is called with a class that is actually derived from
some smobified base class providing is_smob and unsmob, the resulting
check and error message are inaccurate.  This patch uses dynamic_cast to
implement a separate LY_ASSERT_DERIVED_SMOB attaining sensible behavior
in that case.  When no derivation is involved, the performance impact is
minimal.