Issue 4372: convert-ly misses some bflat -> b-flat
In python/convert-rules.py, using '\b' for note name boundaries turns
out to be too audacious since bflat4 is one word.  A proper solution
would be to match the variable wordsyntax and let the match expression
use a replacement function that triggers on the right words.  However,
such a large number of matches would be expensive.  So instead we are
using positive lookahead and lookbehind expressions to get reasonably
good matches.