Issue 4389: Add aliases fflat et al for f-flat et al in bagpipe.ly
While the name conversion of issue 4209 is also done for bagpipe music,
as opposed to "normal" music the respective note names have no short
names and thus can be expected to be used in melodies rather than just
in isolated places like \key commands.

As the change is consequently much more likely to inconvenience bagpipe
players, aliases for the old names are added to bagpipe.ly.