Issue 4401: Remove unnecessary static casts on function member pointers
The commits

6d391dc04f0faa3248d64aa36faf38cf9e4e5fa2 Even more typecasting of member function pointers in smobs.tcc
30203dfe5c45fb9c6255512ba0f8366b322d303f Avoid comparing pointers-to-member of covariant types

were done as hotfixes (apparently without associated tracker issues) to
get the current implementation of smobs to compile with old versions of
GCC in GUB as of Oct 2014.  The attempts to make these versions of GCC
work were ultimately unsuccessful and GUB was upgraded to newer versions
of GCC.

It does not make sense to retain the workarounds that did not really
help.