Issue 3057: misbehavior of \deadNotesOn et al at start of music
User-level commands that do not move to a particular context will tend
to be applied at Bottom context when following any notes or similar
events, but at arbitrary level when being at the start of music.  That's
a bad idea.  This patch moves the head styling overrides to Bottom
level.  While this means that

    \new Staff { \deadNotesOn ...

will only affect the default Voice instead of the whole Staff, one can
still get per-Staff behavior with

    \new Staff \with { \deadNotesOn } { ...

so no functionality is really lost here.