Issue 4498: Provide music functions property{Override,Revert,Set,Unset}
The previous music functions `override' and `revert' (intentionally
colliding with the respective keywords) are removed.  It would have been
too confusing to have collisions for `set' and `unset' as well which are
just too similar to existing Scheme constructs.

A convert-ly rule for the previous calls override/revert is not
attempted: they have not yet been documented and have been available
only for one unstable version.  Whoever happened to use them already is
capable of replacing them.