Let \autochange accept optional arguments for the turning-point and clefs
Issue 4530

With the fix for issue 4465 bassStaffProperties and trebleStaffProperties are gone.
This patch reimplements the functionality to set clefs for the staves and offers the
possibility to set another turning-point apart from middle-C.
This is done with optional arguments for \autochange.

Two regtests are added to reflect these possibilities.
Documentation/notation/keyboards.itely is extended accordingly.

Also inserting the usual remarks about license, etc in autochange.scm.