Issue 4610: ly:round-filled-polygon: optional parameter `extroversion'
`extroversion' specifies where on the specified polygon to draw with the
pen of the given thickness.  Its default of -1 corresponds to the
previous behavior of drawing strictly inside of the given polygon.  An
extroversion value of 0 draws exactly along the given polygon, and a
value of 1 will draw just outside of the given polygon.