Clear fret-diagram- and harp-pedal-input-strings from whitespace
issue 4575

Whitespace-characters are deleted before further processing.

Allows for
\markup \fret-diagram #"s:2;h:5;
6-3;5-5;4-5;3-4;2-3;1-x;"

Also adding errors for typos in fret-diagram with a meaningful message:
\markup \fret-diagram #"s:2;g:r;
6-3;5-5;4-5;3-4;2-3;1-x;"
will return:
fatal error: Unhandled entry in fret-diagram "g:r" in "g:r"

This error would not apply, if something for #\g would be defined
in fret-parse-definition-string somewhere in the future.

Then, the message would be:
fatal error: Unhandled entry in fret-diagram "r" in "g:r"

Something similiar is already in harp-pedals.scm