Scheme function to draw lines based on grob layout
A number of C++ stencil callbacks use Line_interface::line to draw
lines based on line-interface properties defining a particular grob.
This allows control of aspects such as line style (based on the setting
of Grob.style) and fine-tuning of dashed lines through dash-fraction
and dash-period.

This patch gives access to Line_interface::line in Scheme through the
callback ly:line-interface::line.  (The simpler name ly:line was ruled
out in an effort to distinguish it from other functions such as
ly:bracket and ly:circle which do not take a grob argument.)  Users
will be able to create custom stencils with more functionality
(including rewriting certain C++ callbacks--such as Hairpin::print--to
allow for easy modifications without loss of capability.)