Doc: Included/compile.itexi - CG 4.2 - Updated notes on reqs for compiling
Added explicit instructions
for the more common Linux
Distributions as to which
commands to run to download
the correct software packages
to be able to compile LilyPond
and its documentation.

Included instructions for:

* Fedora 22
* Linux Mint (17.1 and LMDE - Betsy)
* OpenSUSE 13.2
* Ubuntu (14.04LTS, 14.10 and 15.04)

These instructions are based
on 2.19.27 and later, but should
also work for any earlier stable
2.18 release.

The lists of separate packages
are now put in a single section
for those distributions not
explicitly listed above. New
information added:

Information regarding Tex Gyre
fonts (required as of 2.19.26
or thereabouts).

More explicit instructions for
TExi2html 1.82 for distros now
using version 5.x.

Added URL for 'autoconf' tools.

Removed information about how to
fix a bug in Bison as this was
referring to a very old version
(1.85) and now just simply
recommend using 2.0 or newer.