Issue 4618: Correct argument handling of Unpure_pure_call::call
When ly:make-unpure-pure-container is called with a single procedure
argument, this procedure is used for both unpure and pure calls.  It
turns out that the calling convention in call_pure_function places the
start/end arguments always in position 2/3 of the call.
Unpure_pure_call::call previously always dropped the last 2 arguments of
a pure call before passing the rest on.  Most calls take exactly 3
arguments (grob start end) where this does not make a difference, but
there may be use cases with a different number of arguments.