Issue 4504/1 add whiteout-style, drop whiteout-box
Introduce whiteout-style grob property, with
options of 'box and 'outline, to be used with
the whiteout grob property, and an equivalent
style property for the whiteout markup command.

Remove the whiteout-box grob property and
markup command, and use the new properties
instead.

Make the box style of whiteout the default
style again, as it was before issue 4418.