Issue 4657: Tweak Rest.staff-position fails with beam
The problem here appears to be that Beam::rest-collision-callback
returned 0 instead of any previously established offset when it
decided not to tamper with the default rest positioning without beam.

This also rewrites Beam::pure-rest-collision-callback (which appears
to have dealt correctly with the offsets here and appears
disconcertingly diverged from the unpure callback) a bit.