Doc - CG: Update information about git-cl
I have moved the information
about how to download and
configure git-cl - that was
previously under the
'Commits and Patches' section -
to its own @node (along side
LilyDev and lily-git) under
the 'quick start' section as
it seemed unnecessarily hard
to find in amongst all the
other information, especially
for new developers.

I've added more explicit steps
for LilyDev users - on how to
update git-cl - and also, more
generally, on how to 'configure'
git cl, especially now that we
are using Allura and need the
extra login/bearer token
requirements.