Issue 4664: Ensure one beam is left in subdivided beams
Issue #4355 introduced the concept that in subdivided beams the
division beam should reflect the length of the subsequent group.

However, with groups equal to or longer than one quarter note this
resulted in *no* beam left.

This commit ensures that at least one beam is returned when a beam is
active and subdivideBeam = ##t