Issue 4702/2: Add conversion function creator make-semitone->pitch
Also contains a function shift-semitone->pitch for moving a given
conversion function to a different key.

The functions work on semitones rather than the whole tones that
LilyPond uses in the accidentals of its pitches since the semitones
are the more natural basis for most applications (including
Midi-related work) as well as for humans: 6 whole notes per octave
sounds a lot less common than 12 semitones.

There are currently no uses of those functions: they are provided as a
convenience.