4704: Improve beam subdivision beaming count
Calculates the remaining length of a beam.
If this is shorter than the regular value at the
subdivision point more beams are used.
(Example: subdivision at 1/8 -> one beam left.
Remaing notes on the beam: 3/32 -> two beams left.)

However, if only one stem is left after the subdivision
this isn't applied in order to have a visual separation.
In this case the default subdivision beam count is used.