Store accidental styles in an alist
This rewrites the way how accidental styles
are stored and accessed.

Previously the two were
mangled in the set-accidental-style
procedure so that adding custom styles
required complete redefinition of
this procedure and copying all the
existing definitions. Now, the actual
specifications of accidental styles
are stored in the 'accidental-styles'
alist, and the set-accidentals-properties
and set-accidental-style procedures
have been replaced by a single procedure,
which reads from that alist and
sets the context properties accordingly.