Doc: LM 2.4.1 and NR 3.3.2 additions and corrections
Issue 4721
NR 3.3.2: nitpicks in lily
example; add a missing space;
better code formatting with
\partcombine remove
printPartCombineTexts setting.

Issue 4717
LM 2.4.1: Replace incorrect
naming rule in LM. This
replaces the incorrect rule
about naming variables with a
simple reference to the
'alphabetic-only' _convention_.
A thorough explanation in the
NR remains to be done.