Issue 4798/3: Define key?, key-list?, key-list-or-music?, key-list-or-symbol?
A key (previously a symbol) can be either a non-negative
integer (obeying the predicate index?) or a symbol.  When used for
overrides/reverts, the first elements of a path may not actually be
numbers but they are flagged as an error by other means: since a full
override path can be the concatenation of several key lists, the leading
elements are not treated differently by the list predicates.

Strictly speaking this should not be the case for key-list-or-music? or
key-list-or-symbol? predicates, but making them behave differently seems
unwarranted.  Arguably key-list-or-symbol? should possibly be
key-list-or-key? instead.