Doc: NR - input.itely: Various improvements
Issue 4677

A variety of nitpicks and
improvements reported by
Federico Bruni

Updated information and
included a new @lilypond
example for \KeepWithTag.

Some spelling corrections
and some simple rewording
of a few paragraphs for
clarity (deleted repeated
words, removed personal
pronouns etc.).

Moved a comment (@c) outside
of an @example so as to not
cause confusion to readers.