Doc: NR - changing-defaults.itely - various improvements
Issue 4761

A variety of nitpicks and
improvements reported by
Federico Bruni

Some spelling corrections
and typos fixed.

The section on \set vs.
\override has been
edited in an attmpt to
simplify the explanation
for the reader.

Fixed a couple of TODOs.

Added a @c that references
an open issue to improve
the example for pure/unpure
containers as the example no
longer illustrates what it
used to as of LP 2.17.x +