Doc: CG - Updates for building the LP Docs using XeLaTex
Issue 4780

Updates to the CG for building
the LP Docs using XeLaTex -
instead of the default pdfTex -
where certain symbols (asterix,
pilcrow, section obelix and
diesis signs) are not printed
in the PDF output.

An @warning{} has been added
with explanation on how to
install texlive-xetex.

Note: This does not mean XeTeX
is a mandatory requirement,
unlike the main TexLive packages
already documented.

This patch also includes
additional small typos
and grammar fixes.