Doc: CG - Re-organize information about 'Patches'
Issue 4794

The information about how to
create, manage, upload and
review patches (as well as
the 'Patch cycle') is in many
different places making it
awkward (and time consuming)
for devs - especially those not
used to the process - to
locate the information.

Some of the information for
'using git' to create patches
does overlap with informeation
on how to submit patches. This
results in a lot of unnecessary
'jumping' about the
documentation.

This patch simply moves
sections - no new information
has been added - although some
of the @node and @*section names
have been changed to help with
re-organization.