Issue 4835: Move Callback_wrapper class to separate file and simplify
Callback_wrapper::make_smob now is only templated on the address of
the (static) trampoline function.  Moving the trampolining to the
actual functions in question makes callbacks quite more versatile and
transparent and obviates the previous need for friend declarations due
to mixing internals of other classes into the Callback_wrapper
definition.