4855: Move old news from news-front to news
As requested in
http://lists.gnu.org/archive/html/bug-lilypond/2016-04/msg00042.html
we should strictly reduce the number of old news items on the
front page. This will probably also make it more straightforward
to think about a new appearance of our entry page.

However, I suggest a change in the old-news display.
Currently "news items" are mixed with release notes, and I think
*if* we want to keep old news available it would be better to
separate these two categories.