#4877: Revert #4747: (Remove (all) uses of is-absolute?)
This reverts

commit f30a8189adbbeefa2103e2c2e194040f66bc2291
Author: Urs Liska <ul@openlilylib.org>
Date:   Tue Jan 19 10:52:33 2016 +0100

    #4747: Remove (all) uses of is-absolute?

====================

As discussion revealed the commit message is not true
and indicates a misconception.
(car ly:input-file-line-char-column a-location) does
*not* always return an absolute path, instead this
depends on how the file path has been passed to LilyPond.

As this commit changed the behaviour of point-and-click
in a somewhat unintentional and unmaintainable way it
is better to revert the commit. Especially as the
original behaviour was not harmful.

========================

The check for absolute paths in  in output-ps.scm
and -svg.scm is unnecessary because
(car ly:input-file-line-char-column a-location)
always returns an absolute, slashified path

Now is-absolute? is not used anymore by LilyPond itself. (reverted from commit f30a8189adbbeefa2103e2c2e194040f66bc2291)