Issue 4931 make deadNote work with other font-settings
In case the current (text-)font does not contain a cross-glyph,
font-name is set temporary to '(), and font-family to feta.
If feta is replaced with another font without glyphs for cross-style
note-heads, it will still fail, though.
The regtest dead-notes.ly is altered accordingly.
Also, ly-syntax is used where possible.
palmMute and harmonic are not affected by this patch and still
work nicely.