Issue 4962: Web: home page: add example image, reduce news to headlines
The example image is the excerpt from Bach BWV 861 from the essay.
Anchor links are used to link to specific examples on the examples
page.  The news entries on the home page are reduced to just a list
of headlines that link to specific entries on the 'Old News' page
(under 'Community').  And that page is renamed to just 'News'.