Issue 4988 straight-flags partly wrong calculated
The angles for straight-flags at down- and up-pointing stems are now
accurately respected.
This will cause some slight changes in the reg-tests.
Also fixing flat-flags, which relied on the old, buggy calculation.
Also doing the same fixes for markup-command note-by-number.
Also fixing an oversight in regtest metronome-mark-formatter.ly,
where the creation of the metronome-markup depends on note-by-number.