Issue 5003 unfoldRepeats can be restricted to certain repeat-types
It now takes an additional, optional argument.
It's a list of symbols representing the repeat-types which should
be unfolded.
Possible settings are percent, tremolo and volta.
The default is an empty list, for which repeated-music is taken,
unfolding all.
A new regtest is added to cover the new feature.
Changes is extended accordingly.

Also correcting two small typos in Documentation/notation/staff.itely