Restrict spread of symbols in church rests
The spacing of symbols in multi-measure church rests varies with
the measure size.  In loose spacing, this can lead to loss of
readability as the symbols are pushed far apart.

This patch introduces the property "max-symbol-separation," which
represents the largest allowed distance between components of a
Kirchenpause.  Distance between symbols will vary with measure size
unless the maximum is exceeded.

Setting the property to a small value will effectively create
constant spacing.  Using a very large value will permit proportional
spacing regardless of measure length (the previous default behavior).

A default value of 8.0 staff spaces is set to allow flexibility in
shorter measures.