Issue 5079/2: with-music-match: lists should match completely
Previously music properties that were matched to a list of wildcards
could contain fewer elements than the list they were matched to,
leading to an error when the wildcard variable was being assigned to.

Now the list lengths have to be matched exactly: a shorter pattern
than the actual list length of the matched property cannot match any
more.