LyricHyphen whiteout
Issue 5033

A proper outline whiteout
would be ideal, but it is
not currently implemented.
The current code fails
because you need a relatively
large whiteout area for a
small object.

A brute force fix is to increase
the number of iterations, but
the result would be a pdf with
a high percentage of whiteout code.

There is the case where a long
melisma produces a single hyphen
event that generates a lot of
hyphens that cross a number of
mensuration lines.

Rectangular or roundedbox whiteout
will produce gaps in all
mensuration lines crossed, even
if there is no collision.

That's not acceptable.