Implement MeasureSpanner
This patch creates a spanner whose ends are oriented to measure
boundaries, a frequent request from users.  The ends of the
spanner may be aligned in various ways to prefatory material.
The spanner may hold text or markup in a centered gap.

The spanner is begun with the command '\startMeasureSpanner'
and ended with '\stopMeasureSpanner'.

Provide two regression tests.