Issue 5116/1: Allow \change to same context type
The \change command no longer refuses to change the parent of a
context when the context itself would be of the proper type to
change to.  This allows changing the parent of a context that
could be nested, like StaffGroup in StaffGroup or NullVoice
in Staff (since NullVoice is also aliased to Staff).