Issue 5127: Implement extra-offset for system positioning
Add the property 'extra-offset to 'line-break-system-details.

It is possible to position systems absolutely using the
properties 'X-offset and 'Y-offset of 'line-break-system-details.
Placement using 'Y-offset, in particular, often requires
considerable trial and error, since it is reckoned from the top
of the usable space.

To remedy this, an 'extra-offset property allows easy movement of
systems relative to where they currently are on the page--whether
there through default or absolute positioning.

This property supports displacements in both axes, though
movement in the Y direction is more generally useful.

System separators will be properly positioned. (Thanks go to Thomas
Morley.)