Issue 5128: Prevent granados.ly image rotation in web.pdf
granados.ly used 'landscape symbol for page settings.
As a result, granados.pdf which was generated from granados.ly
contained `/Rotate 90` that is not supported by XeTeX's PDFs including.
So the image in web.pdf which was generated by XeTeX
was rotated 90 degree counter-clockwise.

This commit makes the page settings does not have 'landscape symbol.
We append `landscape` to the name of the paper size itself.
So granados.ly image rotation in web.pdf is prevented.

See Notation Reference "Setting the paper size".