Issue 5145/2: Allow $... to specify a markup command
This is a partial (though more flexible) substitute for \on-the-fly
since it allows to replace

\on-the-fly #(lambda (layout props args) ...)

with

$(let ((fun (lambda (layout props args) ...)))
   (set! (markup-command-signature fun) (list markup?))
   fun)

Namely: ad-hoc written functions (probably most useful when written
using macros) can be employed as markup commands when preceded with $
as long as they have a suitable markup command signature attached.
This feature is mainly added for consistency with music functions.