Add system-by-system output to SVG backend
This change enables the output of separate files for each system of
music by implementing the `output-classic-framework` function.

Snippet (-dclip-systems) functionality has been added to the backend,
but as calculations for height and width were incorrect, they have been
omited.

For consistency with other backends, the page suffix is reduced from
"-page-~a" to simply "-~a".