Issue 1434: Fix text spacing in SVG
This change makes every SVG contain a style tag with the `white-space`
property set to include all white-space (`pre`). This will cause all
white-space to be applied, and not be ignored.

As SVG files with `-dsvg-woff` set already appended a style tag, the
woff-header now only appends their settings, with a new define to close
the style tag.

The main difficulty arising with this change was with the newline
character that appears after each opening tag. To rectify, a boolean
parameter was added to define(s) on whether to insert a newline for
opening tags, and the appropriate value was passed to all uses of the
define(s).