Issue 5181/1: Parser: Add reverse_music_list function
This function does not just reverse a music list but also integrates
post events into the respective preceding expressions.  If that is not
possible, a warning is printed and they are either dropped or attached
to an empty chord (depending on a call parameter).  If a list consists
only of post-events, it might be returned as a post-event, too.
Again, depending on a parameter.  This is useful for #{ -2 -3 #} and
its ilk.

Also adds add_post_events function.