Limit use of `make-concat-markup` in tempo tag
Issue 5174

Calling `make-concat-markup` when the list of components of the tempo
tag is incomplete causes the printing command to be unnecessarily
fragmented, causing problems, mostly in the SVG backend. This change
combines all components of the metronome mark in one list before
calling `make-concat-markup`.

The make-simple-markup statements are also combined where possible, and
comments have been added where characters can be confused. None have
modified.

This patch provides one part for the fix of bug 5174. Further patches
will be needed to combine text of different styles/fonts or to fix
text spacing in the SVG backend.

Bug: 5174