Issue 5181/2: Allow lone postevents in braced lists
Reattaching post-events to a preceding music variable or music
function call (or expression) is often desired but was unsupported.

This commit caters for the support in braced lists.

The problem with supporting such expressions _everywhere_ is that it
breaks the use of

    c-\single \dynamicUp \f

since then \f is attached to \dynamicUp instead of becoming the second
argument to \single, due to operator precedence.  So reattaching lone
post-events is only done with lower precedence in selected constructs.