Issue 5194: Improve warning for unattached post-event(s)
Insertion of <> was considered too confusing in the given warning:
this isn't mentioned any more.  Additionally, for multiple post events
in a row, a warning is given for every one.  Also, the _type_ of the
post event is printed.

(cherry picked from commit f380b23d379748c34a1319536db04770309d8444)