Adapt lilypond to ghostscript 9.22, change command line options
Issue 5201

We need to adapt lilypond
to ghostscript 9.22 as our
old --bigpdf parameter is
broken by the new version
of gs.

--bigpdf is totally removed.

--pspdfopt=(size/TeX/TeX-GS)
is implemented to allow the
user to select the appropiate
set of -d... parameters for
different purposes.

The patch also contains some
minor fixes of related bugs,
introduces --eps to the
command line interface and
changes the bug-report
message, as post.gmane is
gone.