Issue 5212/1: Allow quoted strings as Scheme arguments to markup commands
After issue 5113, quoted strings and non-quoted strings are
distinguished in the lexer.  This makes it feasible for Scheme-type
arguments of markup commands to accept quoted strings.  Unquoted
strings (token SYMBOL in the lexer) remain out of consideration since
it would be hard to explain why 0.0 is acceptable as a string?
argument but not as number? : in markup mode, most character sequences
end up as "SYMBOL" by default in order to allow typesetting of text
blocks.