Issue 5268: simplify building with a read-only source directory
autogen.sh: If the source directory is not writable, generate the
configure script into the current directory and tell the configure
script the source directory.

This is convenient for those would prefer to build in a container or
virtual machine with a read-only view of source stored on the host.