Issue 2199/1: align outside-staff grobs vertically in staff groups
In any context in which a Vertical_align_engraver creates a
StaffGrouper, it also sees to it that all outside-staff grobs created
in the immediate context become y-axis children of the StaffGrouper.

Any such grobs which are later processed by the after-line-breaking
callback ly:side-position-interface::move_to_extremal_staff are moved
to the extremal staff of the group, not of the entire system.