Issue 5302/3: Let \tweak on event chords manipulate its elements
This makes \tweak on event chords actually get heeded and creates
reasonably predictable behavior by tweaking the rhythmic elements (but
not chord-wide articulations).  It also passes tweaks transparently
through music wrapper expressions (like transpositions, context
changes and other).

To achieve this, it creates and uses a get-tweakable-music function
that will deliver a list of primary music expressions tweaks should be
applied to.

This function is also used in the definition of \once to decide when a
warning for misapplication of \once might be called for.